requirements.txt
alabaster==0.7.13
asgiref==3.6.0
astroid==2.15.2
attrs==22.2.0
Babel==2.12.1
bandit==1.7.5
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
coverage==7.2.2
dill==0.3.6
Django==4.1.7
django-braces==1.15.0
django-environ==0.10.0
django-extensions==3.2.1
django-filter==23.1
django-handyhelpers==0.2.18
django-userextensions==0.2.2
djangoaddicts.codegen @ file:///home/david/projects/django-codegen/dist/djangoaddicts.codegen-0.0.0.tar.gz
djangorestframework==3.13.1
djangorestframework-filters==1.0.0.dev0
docutils==0.19
drf-flex-fields==1.0.0
exceptiongroup==1.1.1
flake9==3.8.3.post2
gitdb==4.0.10
GitPython==3.1.31
hurry.filesize==0.9
idna==3.4
imagesize==1.4.1
iniconfig==2.0.0
isort==5.12.0
Jinja2==3.1.2
lazy-object-proxy==1.9.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mccabe==0.6.1
mdurl==0.1.2
model-bakery==1.10.3
mypy==1.2.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
pbr==5.11.1
platformdirs==3.2.0
pluggy==1.0.0
psutil==5.9.4
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.14.0
pylint==2.17.2
pytest==7.2.2
pytest-cov==4.0.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2023.2
PyYAML==6.0
requests==2.28.2
rich==13.3.3
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
Sphinx==1.8.5
sphinxcontrib-serializinghtml==1.1.5
sphinxcontrib-websupport==1.2.4
sqlparse==0.4.3
stevedore==5.0.0
tomli==2.0.1
tomlkit==0.11.7
typing_extensions==4.5.0
urllib3==1.26.15
wrapt==1.15.0
xlwt==1.3.0
